// packages/architect/src/domain/schema.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface DomainOptions {
  /**
   * Domain name
   */
  name: string;
  /**
   * Add cqrs library
   */
  cqrs?: boolean;
  /**
   * A type to determine if and how to build the library.
   */
  type?: "internal" | "buildable" | "publishable";
  [k: string]: unknown;
}